requirements.txt
google-generativeai
requests
python-dotenv
flask
line-bot-sdk<3.0.0
functions-framework
google-cloud-firestore
google-cloud-storage
google-auth
google-analytics-data
